fix: remove redundant custom raw tabs
